@@ -1,6 +1,6 @@
 {
-  "name": "svip",
-  "version": "0.0.0",
+  "name": "sbox-gui",
+  "version": "1.2.1",
   "main": "app.js",
   "scripts": {
     "ng": "ng",
@@ -13,41 +13,40 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^15.2.0",
-    "@angular/common": "^15.2.0",
-    "@angular/compiler": "^15.2.0",
-    "@angular/core": "^15.2.0",
-    "@angular/forms": "^15.2.0",
-    "@angular/platform-browser": "^15.2.0",
-    "@angular/platform-browser-dynamic": "^15.2.0",
-    "@angular/router": "^15.2.0",
-    "@ng-bootstrap/ng-bootstrap": "^14.1.1",
+    "@angular/animations": "^19.2.14",
+    "@angular/common": "^19.2.14",
+    "@angular/compiler": "^19.2.14",
+    "@angular/core": "^19.2.14",
+    "@angular/forms": "^19.2.14",
+    "@angular/platform-browser": "^19.2.14",
+    "@angular/platform-browser-dynamic": "^19.2.14",
+    "@angular/router": "^19.2.14",
+    "@ng-bootstrap/ng-bootstrap": "^18.0.0",
     "@popperjs/core": "^2.11.6",
-    "bootstrap": "^5.2.3",
-    "electron": "^24.3.0",
-    "electron-packager": "^17.1.1",
+    "bootstrap": "^5.3.7",
+    "fflate": "^0.8.2",
     "file-saver": "^2.0.5",
-    "jszip": "^3.10.1",
-    "rxjs": "~7.8.0",
+    "rxjs": "~7.8.2",
     "tslib": "^2.3.0",
-    "zip-a-folder": "^2.0.0",
-    "zone.js": "~0.12.0"
+    "zip-a-folder": "^3.1.9",
+    "zone.js": "~0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^15.2.0",
-    "@angular/cli": "~15.2.0",
-    "@angular/compiler-cli": "^15.2.0",
-    "@angular/localize": "^15.2.0",
+    "@angular-devkit/build-angular": "^19.2.15",
+    "@angular/cli": "~19.2.15",
+    "@angular/compiler-cli": "^19.2.14",
+    "@angular/localize": "^19.2.14",
     "@types/bootstrap": "^5.2.6",
-    "@types/file-saver": "^2.0.5",
-    "@types/jasmine": "~4.3.0",
-    "electron": "^24.3.0",
-    "jasmine-core": "~4.5.0",
-    "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.2.0",
+    "@types/file-saver": "^2.0.7",
+    "@types/jasmine": "~5.1.9",
+    "electron": "^37.3.0",
+    "@electron/packager": "^18.4.2",
+    "jasmine-core": "~5.9.0",
+    "karma": "~6.4.4",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage": "~2.2.1",
     "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.0.0",
-    "typescript": "~4.9.4"
+    "karma-jasmine-html-reporter": "~2.1.0",
+    "typescript": "~5.8.3"
   }
 }
