@@ -1,23 +1,30 @@
-import { Component } from '@angular/core';
-import { SbomService } from 'src/app/shared/services/sbom.service';
+import {Component} from '@angular/core';
+import {SbomService} from 'src/app/shared/services/sbom.service';
 import filter from '../metrics/models/filters';
-import palettes, { PALETTE, resultStatus } from '../metrics/models/palette';
-import { DownloadService } from 'src/app/shared/services/download.service';
+import palettes, {PALETTE, resultStatus} from '../metrics/models/palette';
+import {DownloadService} from 'src/app/shared/services/download.service';
 
 @Component({
   selector: 'app-comparison',
   templateUrl: './comparison.component.html',
-  styleUrls: ['./comparison.component.css']
+  styleUrls: ['./comparison.component.css'],
+  standalone: false
 })
 export class ComparisonComponent {
   attributes: filter = {};
   palettes = palettes;
   mismatchColor = '#FFA600';
   missingColor = 'var(--warn)';
+
+  constructor(public sbomService: SbomService, public downloadService: DownloadService) {
+  }
+
   private _palette = PALETTE.DEFAULT;
+
   get palette() {
     return this._palette;
   }
+
   set palette(value: PALETTE) {
     this._palette = value;
     this.setColor();
@@ -31,49 +38,20 @@ export class ComparisonComponent {
   }
 
   _resultStatus: resultStatus = {
-    MISMATCH: { shown: true, color: this.mismatchColor},
-    MISSING: { shown: true, color: this.missingColor},
+    MISMATCH: {shown: true, color: this.mismatchColor},
+    MISSING: {shown: true, color: this.missingColor},
   };
-  set resultStatus(value: resultStatus) {
-    this.resultStatus = this.resultStatus;
-  }
+
   get resultStatus() {
     return this._resultStatus;
   }
 
-  constructor(public sbomService: SbomService, public downloadService: DownloadService) { }
+  set resultStatus(value: resultStatus) {
+    this.resultStatus = this.resultStatus;
+  }
 
   GetComparison() {
     return this.sbomService.comparison;
-  }
-
-  protected castToString(value: unknown): string {
-    return value as string;
-  }
-
-  protected formatValue(value: any): string {
-    const formattedValue = JSON.stringify(value, null, 2);
-    const replacedValue = formattedValue.replace(/[[\]{},"]/g, '');
-
-    const capitalizeFirstLetter = (word: string) => {
-        return word.replace(/(\w*)'(\w*)/g, (_, beforeApostrophe, afterApostrophe) => {
-            return beforeApostrophe.toLowerCase() + "'" + afterApostrophe.toLowerCase();
-        });
-    };
-
-    const modifiedValue = replacedValue.replace(/_/g, ' ').replace(/\b\w+\b/g, capitalizeFirstLetter);
-
-    return modifiedValue + '\n';
-  }
-
-  protected castAsAny(value: unknown): any {
-    return value as any;
-  }
-
-  protected removeComment(value: any): any {
-    const valueWithoutComment = { ...value };
-    delete valueWithoutComment.message;
-    return valueWithoutComment;
   }
 
   getFilterType(string: string): boolean {
@@ -96,7 +74,7 @@ export class ComparisonComponent {
     const fileName = 'report.json';
     const reportData = this.GetComparison();
     const reportJson = JSON.stringify(reportData, null, 2);
-    this.downloadService.Download(fileName, new Blob([reportJson], { type: 'application/json' }));
+    this.downloadService.Download(fileName, new Blob([reportJson], {type: 'application/json'}));
   }
 
   /**
@@ -104,5 +82,34 @@ export class ComparisonComponent {
    */
   getAlias(sbom: string) {
     return this.sbomService.getSBOMAliasByID(sbom);
+  }
+
+  protected castToString(value: unknown): string {
+    return value as string;
+  }
+
+  protected formatValue(value: any): string {
+    const formattedValue = JSON.stringify(value, null, 2);
+    const replacedValue = formattedValue.replace(/[[\]{},"]/g, '');
+
+    const capitalizeFirstLetter = (word: string) => {
+      return word.replace(/(\w*)'(\w*)/g, (_, beforeApostrophe, afterApostrophe) => {
+        return beforeApostrophe.toLowerCase() + "'" + afterApostrophe.toLowerCase();
+      });
+    };
+
+    const modifiedValue = replacedValue.replace(/_/g, ' ').replace(/\b\w+\b/g, capitalizeFirstLetter);
+
+    return modifiedValue + '\n';
+  }
+
+  protected castAsAny(value: unknown): any {
+    return value as any;
+  }
+
+  protected removeComment(value: any): any {
+    const valueWithoutComment = {...value};
+    delete valueWithoutComment.message;
+    return valueWithoutComment;
   }
 }
