@@ -1,6 +1,6 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import {ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { DiffFiltersComponent } from './diff-filters.component';
+import {DiffFiltersComponent} from './diff-filters.component';
 
 describe('DiffFiltersComponent', () => {
   let component: DiffFiltersComponent;
@@ -8,9 +8,9 @@ describe('DiffFiltersComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ DiffFiltersComponent ]
+      declarations: [DiffFiltersComponent]
     })
-    .compileComponents();
+      .compileComponents();
 
     fixture = TestBed.createComponent(DiffFiltersComponent);
     component = fixture.componentInstance;
