@@ -1,1 +1,3 @@
-export default interface filters  {[processorName: string]: { color: string; shown: boolean }};
+export default interface filters {
+  [processorName: string]: { color: string; shown: boolean }
+};
