@@ -1,15 +1,17 @@
-import { Component, Input, OnInit} from '@angular/core';
+import {Component, Input, OnInit} from '@angular/core';
 
 @Component({
   selector: 'app-button',
   templateUrl: './button.component.html',
-  styleUrls: ['./button.component.css']
+  styleUrls: ['./button.component.css'],
+  standalone: false
 })
 export class ButtonComponent implements OnInit {
   @Input() text: string = '';
   @Input() icon: string = "";
 
-  constructor() { }
+  constructor() {
+  }
 
   ngOnInit() {
   }
