@@ -1,11 +1,11 @@
 enum color {
-    red = "var(--red)",
-    orange = "var(--orange)",
-    yellow = "var(--yellow)",
-    green = "var(--green)",
-    teal = "var(--teal)",
-    blue = "var(--blue)",
-    pink = "var(--pink)",
-  }
-  
-  export default color;
+  red = "var(--red)",
+  orange = "var(--orange)",
+  yellow = "var(--yellow)",
+  green = "var(--green)",
+  teal = "var(--teal)",
+  blue = "var(--blue)",
+  pink = "var(--pink)",
+}
+
+export default color;
