@@ -1,11 +1,11 @@
 enum shape {
-    square = 'square',
-    circle = 'circle',
-    hexagon = 'hexagon',
-    pentagon = 'pentagon',
-    triangle = 'navigation',
-    star = 'star',
-    eightStar = 'brightness_empty',
-  }
-  
-  export default shape;
+  square = 'square',
+  circle = 'circle',
+  hexagon = 'hexagon',
+  pentagon = 'pentagon',
+  triangle = 'navigation',
+  star = 'star',
+  eightStar = 'brightness_empty',
+}
+
+export default shape;
