@@ -1,13 +1,14 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
+import {Component, EventEmitter, Input, Output} from '@angular/core';
 
 @Component({
   selector: 'app-multiselect-dropdown',
   templateUrl: './multiselect-dropdown.component.html',
-  styleUrls: ['./multiselect-dropdown.component.css']
+  styleUrls: ['./multiselect-dropdown.component.css'],
+  standalone: false
 })
 export class MultiselectDropdownComponent {
   @Input() title: string = 'Select Options';
-  @Input() options: { [name: string]: boolean} = {};
+  @Input() options: { [name: string]: boolean } = {};
   @Output() checkboxChange: EventEmitter<{ name: string, value: boolean }> = new EventEmitter();
 
   opened: boolean = false;
@@ -32,6 +33,6 @@ export class MultiselectDropdownComponent {
     // Get the checkbox value
     const checkboxValue = event.target.checked;
 
-    this.checkboxChange.emit({ name: value, value: checkboxValue });
+    this.checkboxChange.emit({name: value, value: checkboxValue});
   }
 }
