@@ -1,15 +1,17 @@
-import { Component } from '@angular/core';
-import { SVIPService } from '../../services/SVIP.service';
-import { SbomService } from '../../services/sbom.service';
-import { UploadComponent } from 'src/app/features/upload/upload.component';
+import {Component} from '@angular/core';
+import {SVIPService} from '../../services/SVIP.service';
+import {SbomService} from '../../services/sbom.service';
+import {UploadComponent} from 'src/app/features/upload/upload.component';
 
 @Component({
   selector: 'app-toggle',
   templateUrl: './toggle.component.html',
-  styleUrls: ['./toggle.component.css']
+  styleUrls: ['./toggle.component.css'],
+  standalone: false
 })
 export class ToggleComponent {
   upload: boolean = false;
+
   constructor(private svipService: SVIPService, private sbomService: SbomService, private uploadComp: UploadComponent) {
   }
 
