@@ -1,15 +1,14 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
-import { RoutingService } from 'src/app/shared/services/routing.service';
-import { SbomService } from 'src/app/shared/services/sbom.service';
+import {Component, EventEmitter, Input, Output} from '@angular/core';
+import {RoutingService} from 'src/app/shared/services/routing.service';
+import {SbomService} from 'src/app/shared/services/sbom.service';
 
 @Component({
   selector: 'app-convert-modal',
   templateUrl: './convert-modal.component.html',
-  styleUrls: ['./convert-modal.component.css']
+  styleUrls: ['./convert-modal.component.css'],
+  standalone: false
 })
 export class ConvertModalComponent {
-  constructor(private sbomService: SbomService, public routing: RoutingService) {}
-
   public convertOptions: {
     schema: string;
     format: string;
@@ -19,11 +18,12 @@ export class ConvertModalComponent {
     format: '',
     overwrite: true,
   };
-
-  @Input() choices!: {[key: string]: string[]};
-
+  @Input() choices!: { [key: string]: string[] };
   @Input() opened: boolean = false;
   @Output() close = new EventEmitter<Boolean>();
+
+  constructor(private sbomService: SbomService, public routing: RoutingService) {
+  }
 
   GetSelected() {
     const checkboxes = document.querySelectorAll('.sbom-checkbox');
