@@ -1,17 +1,19 @@
-import { Component, Input, Output, EventEmitter } from '@angular/core';
-import { PAGES, RoutingService } from 'src/app/shared/services/routing.service';
-import { SbomService } from 'src/app/shared/services/sbom.service';
+import {Component, EventEmitter, Input, Output} from '@angular/core';
+import {PAGES, RoutingService} from 'src/app/shared/services/routing.service';
+import {SbomService} from 'src/app/shared/services/sbom.service';
 
 @Component({
   selector: 'app-delete-modal',
   templateUrl: './delete-modal.component.html',
-  styleUrls: ['./delete-modal.component.css']
+  styleUrls: ['./delete-modal.component.css'],
+  standalone: false
 })
 export class DeleteModalComponent {
-  constructor(private sbomService: SbomService, public routing: RoutingService) {}
-
   @Input() opened: boolean = false;
   @Output() close = new EventEmitter<Boolean>();
+
+  constructor(private sbomService: SbomService, public routing: RoutingService) {
+  }
 
   GetSelected() {
     const checkboxes = document.querySelectorAll('.sbom-checkbox');
@@ -27,11 +29,11 @@ export class DeleteModalComponent {
     return selected;
   }
 
-   /**
+  /**
    * Removes file from uploaded files
    * @param file file to remove
    */
-   RemoveFile(file: string) {
+  RemoveFile(file: string) {
     if (this.routing.GetPage() === PAGES.VIEW && this.routing.data === file) {
       this.routing.SetPage(PAGES.NONE);
       this.routing.data = undefined;
