@@ -1,13 +1,14 @@
-import { Component, Input, Output, EventEmitter, OnInit } from '@angular/core';
-import { SVIPService } from 'src/app/shared/services/SVIP.service';
-import { Subject } from 'rxjs';
-import { SbomService } from 'src/app/shared/services/sbom.service';
-import { ToastService } from 'src/app/shared/services/toast.service';
+import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
+import {SVIPService} from 'src/app/shared/services/SVIP.service';
+import {Subject} from 'rxjs';
+import {SbomService} from 'src/app/shared/services/sbom.service';
+import {ToastService} from 'src/app/shared/services/toast.service';
 
 @Component({
   selector: 'app-generate-modal',
   templateUrl: './generate-modal.component.html',
-  styleUrls: ['./generate-modal.component.css']
+  styleUrls: ['./generate-modal.component.css'],
+  standalone: false
 })
 export class GenerateModalComponent implements OnInit {
   public options: {
@@ -22,56 +23,56 @@ export class GenerateModalComponent implements OnInit {
     type: '',
   };
 
-  public choices: {[key: string]: string[]} = {
+  public choices: { [key: string]: string[] } = {
     'CDX14': ['JSON', 'XML'],
     'SPDX23': ['TAGVALUE', 'JSON'],
   }
 
   public types: string[] = ['OSI', 'PARSERS'];
 
-  public osiTools: {[name: string]: boolean} = {};
+  public osiTools: { [name: string]: boolean } = {};
 
   @Input() opened: boolean = false;
   @Output() close = new EventEmitter<Boolean>();
-  private openedSubject = new Subject<boolean>();
-
   public status: GenerationStatus = GenerationStatus.NULL;
   public zippedFileData: any;
+  private openedSubject = new Subject<boolean>();
 
-  constructor(private service: SVIPService, private sbomService: SbomService, private toast: ToastService) {}
+  constructor(private service: SVIPService, private sbomService: SbomService, private toast: ToastService) {
+  }
 
   ngOnInit(): void {
     this.openedSubject.subscribe((value) => {
-      if(!value)
+      if (!value)
         return;
 
-        this.zippedFileData = undefined;
-        this.status = GenerationStatus.GENERATING;
+      this.zippedFileData = undefined;
+      this.status = GenerationStatus.GENERATING;
 
-        this.service.getProjectDirectory().then((result) => {
-          this.status = GenerationStatus.ZIPPING;
+      this.service.getProjectDirectory().then((result) => {
+        this.status = GenerationStatus.ZIPPING;
 
-          this.service.zipFileDirectory(result).then((data) => {
+        this.service.zipFileDirectory(result).then((data) => {
 
-            //TODO: Prompt user beforehand on OSI or Parsers so don't need to upload project if don't have to OR the backend should be reworked for parsers
-            this.service.uploadProject(data, 'osi').then((tools: any) => {
+          //TODO: Prompt user beforehand on OSI or Parsers so don't need to upload project if don't have to OR the backend should be reworked for parsers
+          this.service.uploadProject(data, 'osi').then((tools: any) => {
 
-              tools.forEach((tool: any) => {
-                this.osiTools[tool] = true;
-              })
-
-              this.zippedFileData = data;
-              this.status = GenerationStatus.PROJECT_INFO;
-
-            }).catch((error) => {
-              this.Close();
+            tools.forEach((tool: any) => {
+              this.osiTools[tool] = true;
             })
+
+            this.zippedFileData = data;
+            this.status = GenerationStatus.PROJECT_INFO;
+
           }).catch((error) => {
             this.Close();
           })
         }).catch((error) => {
           this.Close();
         })
+      }).catch((error) => {
+        this.Close();
+      })
     });
   }
 
@@ -98,12 +99,12 @@ export class GenerateModalComponent implements OnInit {
       this.options.format,
       this.options.type,
       tools).then((data: any) => {
-        this.sbomService.addSBOMbyID(data);
-        this.Close();
-      }).catch(() => {
-        this.toast.showErrorToast("SBOM Generation", "Failed");
-        this.Close();
-      })
+      this.sbomService.addSBOMbyID(data);
+      this.Close();
+    }).catch(() => {
+      this.toast.showErrorToast("SBOM Generation", "Failed");
+      this.Close();
+    })
 
 
   }
