@@ -1,6 +1,6 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import {ComponentFixture, TestBed} from '@angular/core/testing';
 
-import { RepairModalComponent } from './repair-modal.component';
+import {RepairModalComponent} from './repair-modal.component';
 
 describe('RepairModalComponent', () => {
   let component: RepairModalComponent;
@@ -8,9 +8,9 @@ describe('RepairModalComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      declarations: [ RepairModalComponent ]
+      declarations: [RepairModalComponent]
     })
-    .compileComponents();
+      .compileComponents();
 
     fixture = TestBed.createComponent(RepairModalComponent);
     component = fixture.componentInstance;
