@@ -1,16 +1,16 @@
-import { Injectable } from '@angular/core';
-import * as JSZip from 'jszip';
-import { saveAs } from 'file-saver';
+import {Injectable} from '@angular/core';
+import {strToU8, zipSync} from 'fflate';
 
 @Injectable({
   providedIn: 'root'
 })
 export class DownloadService {
 
-  constructor() { }
+  constructor() {
+  }
 
   Download(fileName: string, content: Blob) {
-    const blob = new Blob([content], { type: 'text/plain' });
+    const blob = new Blob([content], {type: 'text/plain'});
     const url = window.URL.createObjectURL(blob);
 
     const anchor = document.createElement('a');
@@ -21,20 +21,27 @@ export class DownloadService {
     window.URL.revokeObjectURL(url);
   }
 
-  DownloadAsZip(files: {[key: string]: string}) {
-    const zip = new JSZip();
+  DownloadAsZip(files: { [key: string]: string }) {
+    const zipData: Record<string, Uint8Array> = {};
 
     let keys = Object.keys(files);
 
-    for(let i = 0; i < keys.length; i++) {
+    for (let i = 0; i < keys.length; i++) {
       let key = keys[i];
       let contents = files[key];
 
-      zip.file(key, contents);
+      zipData[key] = strToU8(contents);
     }
 
-    zip.generateAsync({ type: 'blob' }).then((content) => {
-    saveAs(content, 'files.zip');
-  });
+    const content = new Blob([zipSync(zipData)], {type: 'application/zip'});
+    this.#saveAs(content, 'files.zip');
+  }
+
+  #saveAs(blob: Blob, filename: string) {
+    const link = document.createElement('a');
+    link.href = URL.createObjectURL(blob);
+    link.download = filename;
+    link.click();
+    URL.revokeObjectURL(link.href);
   }
 }
