@@ -1,7 +1,7 @@
-import { Injectable } from '@angular/core';
-import { Observable, Subject } from 'rxjs';
-import { EventTypes } from '../models/event-types';
-import { ToastEvent } from '../models/toast-event';
+import {Injectable} from '@angular/core';
+import {Observable, Subject} from 'rxjs';
+import {EventTypes} from '../models/event-types';
+import {ToastEvent} from '../models/toast-event';
 
 @Injectable({
   providedIn: 'root',
@@ -13,7 +13,7 @@ export class ToastService {
   constructor() {
     this.toastEvents = this._toastEvents.asObservable();
   }
-  
+
   showWarningToast(title: string, message: string) {
     this._toastEvents.next({
       message,
